add user CRUD, RBAC

+++ app/Http/Controllers/RoleController.php
@@ -4,6 +4,7 @@ namespace App\Http\Controllers;
 
 use App\Http\Requests\StoreRoleRequest;
 use App\Http\Requests\UpdateRoleRequest;
+use App\Models\Permission;
 use App\Models\Role;
 use Illuminate\Http\Request;
 
@@ -13,7 +14,9 @@ class RoleController extends Controller
     {
         $roles = Role::paginate(10);
 
-        return view('roles.index', compact('roles'));
+        $user = auth()->user();
+
+        return view('roles.index', compact('roles', 'user'));
     }
 
     public function store(StoreRoleRequest $request)
@@ -23,16 +26,41 @@ class RoleController extends Controller
         return redirect()->route('web.roles.index');
     }
 
+    public function show(Role $role)
+    {
+        if ($role->id === 1) {
+            return redirect()->route('web.roles.index');
+        }
+
+        $user = auth()->user();
+
+        $permissions = Permission::all();
+
+        return view('roles.show', compact('role', 'user', 'permissions'));
+    }
+
     public function update(UpdateRoleRequest $request, Role $role)
     {
+        if ($role->id === 1) {
+            return redirect()->route('web.roles.index');
+        }
+
         $role->update($request->validated());
 
-        return redirect()->route('web.roles.index');
+        if ($request->has('selected')) {
+            $selectedPermIds = $request->input('selected', []);
+            $selectedPermissions = Permission::whereIn('id', $selectedPermIds)->get();
+            $role->permissions()->sync($selectedPermissions);
+        }
+
+        return redirect()->route('web.roles.show', $role->id);
     }
 
     public function destroy(Role $role)
     {
-        $role->delete();
+        if ($role->id !== 1) {
+            $role->delete();
+        }
 
         return redirect()->route('web.roles.index');
     }
